Work with data move to FindTre ang Users classes
Building a prefix tree and searching through it are wrapped in a class PrefixTree
Add class `User`
Add tests with real data for class . Delete old tests.
All tests move to `__test__` directory
Server is now use the new classes. On the front the order of data processing is changed.